Move merge_deploy.yaml from examples to workflows
This workflow is an example of how you can call the Deploy Flex workflow in response to events. This example will deploy whenever a commit is pushed to the main or develop branches, using the 'production' environment when pushing to the main branch, or 'dev' environment when pushing to the develop branch.

To use this workflow, move it to the .github/workflows directory and make the necessary adjustments for your use case.